Add create and delete functionality for tickets

+++ src/main/java/com/othr/rentopia/controller/TicketController.java
@@ -3,6 +3,7 @@ package com.othr.rentopia.controller;
 import com.othr.rentopia.model.Device;
 import com.othr.rentopia.model.Ticket;
 import com.othr.rentopia.service.TicketServiceImpl;
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,15 +28,42 @@ public class TicketController {
 
     @DeleteMapping("{ticketId}")
     public ResponseEntity<String> deleteTicket(@PathVariable("ticketId") Long ticketId){
-        ticketService.deleteTicket(ticketId);
+        if(ticketId != null && ticketId >= 0){
+            try{
+                ticketService.deleteTicket(ticketId);
+            } catch(Exception e){
+                System.out.println("Deleting ticket threw excption : "+ e.getMessage());
+            }
+        }
+
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping()
     public ResponseEntity<String> createTicket(@RequestBody String ticketInfo){
-        //todo parse ticketInfo
+        JSONObject request = new JSONObject(ticketInfo);
+
         Ticket newTicket = new Ticket();
-        ticketService.saveTicket(newTicket);
+
+        if(request.get("id") != null){
+            //todo ?
+        }
+
+        newTicket.setStatus(Ticket.Status.Open);
+        newTicket.setTitle((String)request.get("title"));
+        newTicket.setDetails((String)request.get("details"));
+        Integer ownerId = (Integer) request.get("ownerId");
+        newTicket.setOwnerID(ownerId.longValue());
+
+        String category = (String) request.get("category");
+        newTicket.setCategory(Ticket.Category.valueOf(category.toLowerCase()));
+
+        try{
+            ticketService.saveTicket(newTicket);
+        } catch(Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
         return new ResponseEntity<>(HttpStatus.OK);
     }
 }
